DVT-#209 맵각코 페이지 검색 바를 개별 컴포넌트로 분리
